fiz algumas alterações nas models e nas controllers

+++ Controllers/CategoryController.cs
@@ -17,16 +17,18 @@ namespace projeto_web.Controllers
         [HttpGet]
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Categories.ToListAsync());
+            return View(await _context.Categories.AsNoTracking().ToListAsync());
         }
 
         [HttpGet]
-        public IActionResult create(){
+        public IActionResult create()
+        {
             return View();
         }
 
         [HttpPost]
-        public async Task<IActionResult> store(Category data){
+        public async Task<IActionResult> store(Category data)
+        {
            if(ModelState.IsValid)
            {
                 await _context.Categories.AddAsync(data);
@@ -38,12 +40,18 @@ namespace projeto_web.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> edit(int id){
+        public async Task<IActionResult> edit(int? id)
+        {
+            if(id == null){
+                return NotFound();
+            }
+
             return View(await _context.Categories.FindAsync(id));
         }
 
         [HttpPost]
-        public async Task<IActionResult> update(Category data){
+        public async Task<IActionResult> update(Category data)
+        {
            if(ModelState.IsValid)
            {
                 _context.Categories.Update(data);
@@ -55,11 +63,16 @@ namespace projeto_web.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> destroy(int id){
-            
+        public async Task<IActionResult> destroy(int id)
+        {
             Category category = await _context.Categories.FindAsync(id);
-            _context.Categories.Remove(category);
-            await _context.SaveChangesAsync();
+            
+            if(category != null)
+            {
+                _context.Categories.Remove(category);
+                await _context.SaveChangesAsync();
+            }
+            
             return RedirectToAction(nameof(Index));
         }  
     }
